Add REPL ns & supernatural summaries text.
Add the `clj-markov.repl` ns, which refers the necessary functions to go
from file to markov chain output, and gives an example of doing this
with the supernatural plot summaries text.

Add `texts/supernatural.txt`, which contains plot summaries for the
first three seasons of the long-running American cult television show
Supernatural. The summaries are from Wikipedia and included under the
terms of the Creative Commons Attribution Share Alike license (see
CC-LICENSE for full details).